Fixed type hint warnings in level_scene.py

+++ src/gui/position.py
@@ -7,4 +7,4 @@ from typing import Union
 
 import pygame
 
-Position = Union[pygame.Vector2, tuple[int, int]]
+Position = Union[pygame.Vector2, tuple[float, float]]
